requirements.txt
django
djangorestframework
mysqlclient
python-decouple
instructlab
PyJWT
dateparser
